pyproject.toml
[project]
name = "ai-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.3.0",
    "google-adk>=1.3.0",
    "ipykernel>=6.29.5",
    "markdown>=3.8",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "prophet>=1.1.7",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
]
